Bifrost UI removed and moved editor toolbar as props from host
