--- src/DraftToolbar/DraftLineSeparator.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-
-interface IDraftLineSeparator {
-    className?: string;
-}
-export const DraftLineSeparator = (props: IDraftLineSeparator) => {
-    const { className } = props;
-
-    return <div className={`line-separator ${className || ''}`} />;
-};
