+++ src/Service/UIconstants.ts
@@ -1,5 +1,4 @@
 export const TB = {
     FONT_BOLD: 'fontBold-b',
     FONT_FAMILY: 'fontSizeIncrement-b',
-
 }
